==> 01_data_preprocessing.ipynb <==
## 01 - Data Preprocessing
In this notebook, we will read in and pre-process data in train and dev set

import pandas as pd

path = "./data/MultiWOZ/"

utt_train = "./data/MultiWOZ/WOZ_train_utt.txt"
ans_train = "./data/MultiWOZ/WOZ_train_ans.txt"

# read in train set utterance
lines = []
with open(utt_train, encoding="utf-8") as f:
    lines = f.readlines()
    print("Length =", len(lines))

df = pd.DataFrame({"text": lines})

# apply data cleaning
df["text"] = df["text"].apply(lambda text: text.strip())
df.head()

df.iloc[110]["text"]

# read in answers
lines = []
with open(ans_train, encoding="utf-8") as f:
    lines = f.readlines()
    print("Length =", len(lines))

# df["answer"]
# order = hotel-area / hotel-internet

df["answer"] = lines
df["answer"] = df["answer"].apply(lambda text: text.strip())
df.head()

df.iloc[110]["answer"]

# split the slots
df["answer"] = df["answer"].apply(lambda text: text.split("|"))
df.head()

def list2dict(item_slots):
    """ 
    convert slot lists in format A=B to dict A:B
    """
    result = dict()
    # skip the first item (intent)
    for slot in item_slots[1:]:
        # print(slot)
        key, value = slot.split("=")
        result[key] = value
    
    return result

list2dict(["find_restaurant", "restaurant-food=asian oriental", "restaurant-pricerange=expensive"])

# get the intent
df["intent"] = df["answer"].apply(lambda item:item[0])
df.head()

# build the slots into dictionary format
df["slots"] = df["answer"].apply(list2dict)
df.head()

# check percentage for 'dontcare' answer
w = 'dontcare'
sum(df["slots"].apply(lambda slots: w in slots.values())) * 100 / len(df)

output_file = "./data/train.csv"
print("Writing to ", output_file)
df.to_csv(output_file, index=False)
print("Completed")

# analyze the slots
slots = set() # empty set
restauant_slots = set()
hotel_slots = set()

n = len(df)
max_row = n # set to n for full set

print("Building slot sets...")
for item in df["answer"][:max_row]:
    # print(item)
    # item_set = set(item)
    # print(item_set)
    slots.update(item)
    restauant_slots.update([slot 
                            for slot in item
                            if slot.startswith("restaurant")
                           ])
    
    hotel_slots.update([slot 
                        for slot in item
                        if slot.startswith("hotel")
                        ])

print("Done.")
print("Slot count =", len(slots))
# print(slots)
print("Restaurant slots =", len(restauant_slots))
print("Hotel slots =", len(hotel_slots))

# finding unique slots keys
from collections import defaultdict

restaurant_dict = defaultdict(list)
for item in restauant_slots:
    key, value = item.split("=")
    restaurant_dict[key].append(value)

print("Keys =", restaurant_dict.keys())


for k in restaurant_dict.keys():
    print(k, restaurant_dict[k])
    print("-------")

hotel_dict = defaultdict(list)
for item in hotel_slots:
    key, value = item.split("=")
    hotel_dict[key].append(value)

print("Keys =", hotel_dict.keys())

for k in hotel_dict.keys():
    print(k, hotel_dict[k])
    print("-------")

# process dev set

utt_dev = "./data/MultiWOZ/WOZ_dev_utt.txt"
ans_dev = "./data/MultiWOZ/WOZ_dev_ans.txt"

# read in train set utterance
lines_utt = []
with open(utt_dev, encoding="utf-8") as f:
    lines_utt = f.readlines()
    print("Length =", len(lines_utt))

lines_ans = []
with open(ans_dev, encoding="utf-8") as f:
    lines_ans = f.readlines()
    print("Length =", len(lines_ans))

dev_df = pd.DataFrame({"text": lines_utt,
                       "answer_raw": lines_ans
                      })

# apply data cleaning
dev_df["text"] = dev_df["text"].apply(lambda text: text.strip())
dev_df["answer_raw"] = dev_df["answer_raw"].apply(lambda text: text.strip())

dev_df.head()

# split the slots
dev_df["answer"] = dev_df["answer_raw"].apply(lambda text: text.split("|"))
# get the intent
dev_df["intent"] = dev_df["answer"].apply(lambda item:item[0])
# build the slots into dictionary format
dev_df["slots"] = dev_df["answer"].apply(list2dict)

dev_df.head()

output_file = "./data/dev.csv"
print("Writing to ", output_file)
dev_df.to_csv(output_file, index=False)
print("Completed")

output_file = "./data/dev.json"
print("Writing to ", output_file)
dev_df.to_json(output_file, orient="records")
print("Completed")
