==> 02_intent_classifier.ipynb <==
## 02 - Intent Classifier
In this notebook, we will classify the intent to "find-hotel" or "find-restaurant" using lexcial approach (most indicative words associated with each intent)

import pandas as pd

train_file = "./data/train.csv"
train_df = pd.read_csv(train_file)
train_df.head()

train_df["text_lower"] = train_df["text"].str.lower()
train_df.head()

from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score

vectorizer = TfidfVectorizer()
x_train = train_df["text_lower"]
x_train_tfidf = vectorizer.fit_transform(x_train)
print(x_train_tfidf.shape)

# train using logistic regression

clf = LogisticRegression(random_state=42)
y_train = train_df["intent"]
print(y_train.shape)
clf.fit(x_train_tfidf, y_train)
print("Completed")

y_pred = clf.predict(x_train_tfidf)
print("Train score =", clf.score(x_train_tfidf, y_train))
print("Accuracy score =", accuracy_score(y_train, y_pred)) # should be the same

# test on dev set
dev_utt = "./data/MultiWOZ/WOZ_dev_utt.txt"
# read in dev set utterance
lines = []
with open(dev_utt, encoding="utf-8") as f:
    lines = f.readlines()
    print("Length =", len(lines))

dev_df = pd.DataFrame({"text":lines})
dev_df.head()

# read dev intent
dev_ans = "./data/MultiWOZ/WOZ_dev_ans.txt"
# read in dev set utterance
lines = []
with open(dev_ans, encoding="utf-8") as f:
    lines = f.readlines()
    print("Length =", len(lines))

dev_df["answer_raw"] = lines
dev_df.head()

# test extract intent from answer raw
s = dev_df.iloc[0]["answer_raw"]
print(s)
print(s.split("|")[0])

dev_df["intent"] = dev_df["answer_raw"].apply(lambda s:s.split("|")[0].strip())
dev_df.head()

x_dev = dev_df["text"]
x_dev_tfidf = vectorizer.transform(x_dev)
print(x_dev_tfidf.shape)
y_dev = dev_df["intent"]
dev_score = clf.score(x_dev_tfidf, y_dev)
dev_pred = clf.predict(x_dev_tfidf)
print("Dev accuracy =", accuracy_score(y_dev, dev_pred) * 100)
print("Dev accuracy =", dev_score * 100)

from collections import Counter
Counter(dev_pred)

# find where it predicted wrong
dev_df["predicted"] = dev_pred
wrong_df = dev_df.query("predicted != intent")
wrong_df

#### Predict intent on test set

# read in data
test_utt = "./data/MultiWOZ/WOZ_test_utt.txt"
# read in dev set utterance
lines = []
with open(test_utt, encoding="utf-8") as f:
    lines = f.readlines()
    print("Length =", len(lines))

test_df = pd.DataFrame({"text":lines})
print(len(test_df))
test_df.head()

# pre-processing
test_df["text_lower"] = test_df["text"].str.lower()
test_df.head()

# predict
x_test = test_df["text_lower"]
x_test_tfidf = vectorizer.transform(x_test)
print(x_test_tfidf.shape)
# y_test = test_df["intent"]
# dev_score = clf.score(x_dev_tfidf, y_dev)
test_pred = clf.predict(x_test_tfidf)
# print("Dev accuracy =", accuracy_score(y_dev, dev_pred) * 100)
# print("Dev accuracy =", dev_score * 100)

test_df["predicted"] = test_pred
test_df

# write to output
test_output_fn = "./data/test_step1.csv"
test_df.to_csv(test_output_fn)
print("Output completed")
