==> 03_hotel_data_extractor.ipynb <==
## 03 - Hotel Information Extraction
In this notebook, we try using simple classifier to extract some information on hotel like price-range, location, etc

import pandas as pd

train_file = "./data/train.csv"
train_df = pd.read_csv(train_file)
train_df.head()

train_df["text_lower"] = train_df["text"].str.lower()

# convert slots to json objects
train_df["slots"] = train_df["slots"].apply(lambda s: eval(s))
train_df.head()

# filter to hotel only
category = 'find_hotel'
train_df = train_df.query("intent==@category")
print(len(train_df))
train_df.head()

# extract slot price-range
k = 60
# item = eval(train_df.iloc[k]["slots"])
item = train_df.iloc[k]["slots"]
print(item)
item.keys()

# kw = 'hotel-pricerange'
kw = 'hotel-stars'
train_df["target"] = train_df["slots"].apply(lambda slots:slots.get(kw, ""))
train_df.head()

from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score

vectorizer = TfidfVectorizer()
x_train = train_df["text_lower"]
x_train_tfidf = vectorizer.fit_transform(x_train)
print(x_train_tfidf.shape)

# train using logistic regression

# kw = 'hotel-internet'
clf = LogisticRegression(random_state=42)
y_train = train_df["target"]
print(y_train.shape)
clf.fit(x_train_tfidf, y_train)
print("Completed")

y_pred = clf.predict(x_train_tfidf)
print("Train score =", clf.score(x_train_tfidf, y_train))
print("Accuracy score =", accuracy_score(y_train, y_pred)) # should be the same

# test on dev set
dev_filename = "./data/dev.csv"
dev_df = pd.read_csv(dev_filename)
print(dev_df.columns)
dev_df.head()

# apply some pre-processing
dev_df["text_lower"] = dev_df["text"].str.lower()
# convert slots to json objects
dev_df["slots"] = dev_df["slots"].apply(lambda s: eval(s))
dev_df.head()

# extract price-range
# kw = 'hotel-pricerange'

dev_df["target"] = dev_df["slots"].apply(lambda slots:slots.get(kw, ""))
dev_df.head()

x_dev = dev_df["text"]
x_dev_tfidf = vectorizer.transform(x_dev)
print(x_dev_tfidf.shape)
y_dev = dev_df["target"]
dev_score = clf.score(x_dev_tfidf, y_dev)
dev_pred = clf.predict(x_dev_tfidf)
print("Slot prediction =", kw)
print("Dev accuracy =", accuracy_score(y_dev, dev_pred) * 100)
print("Dev accuracy =", dev_score * 100)

from collections import Counter
Counter(dev_pred)

# find where it predicted wrong
dev_df["predicted"] = dev_pred
wrong_df = dev_df.query("predicted != target")
print(f"Number of wrong counts = {len(wrong_df)} / {len(dev_df)}")
wrong_df

# find where it predicted correctly, and not null
non_null_count = len(dev_df.query("target != ''"))
dev_df["predicted"] = dev_pred
correct_df = dev_df.query("predicted != '' and predicted == target")
print(f"Number of correct counts = {len(correct_df)} / {non_null_count}")
correct_df
