==> 07_three_slots.ipynb <==
## Step 3 - Three slots (hotel-name, restaurant-food, restaurant-name)
In this notebook, we use Logistic Regression binary classifier to determine if hotel-name, restaurant-food, restaurant-name exists.  
Then, we implement QA model to identify the information carried.

# from sklearn.feature_extraction.text import TfidfVectorizer

from transformers import AutoTokenizer
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score
from transformers import AutoModelForQuestionAnswering, AutoTokenizer, pipeline


import pandas as pd

slot_names = ['hotel-name', 
              'restaurant-food', 
              'restaurant-name',
             ]

train_file = "./data/train.csv"
train_df = pd.read_csv(train_file)

train_df.head()

train_df["text_lower"] = train_df["text"].str.lower()

# convert slots to json objects
train_df["slots"] = train_df["slots"].apply(lambda s: eval(s))
train_df.head()

# filter to hotel only
# train_df = train_df.query("intent==@category")
print(len(train_df))
# train_df.reset_index(inplace=True)
# train_df.drop(columns=["index"], inplace=True)
train_df.head()

# create sentence transformer embeddings
from sentence_transformers import SentenceTransformer

model = SentenceTransformer("sentence-transformers/LaBSE")

x_train = train_df["text_lower"]
x_train = list(train_df["text_lower"])

x_train_labse = model.encode(x_train)

print(x_train_labse.shape)

# train using logistic regression


def transform_to_binary(slots, slot):
    if slot in list(slots.keys()):
        res = "_"
    else:
        res = ""
    return res

# kw = 'hotel-internet'
# create a master classifier, which contains classifier for each slot
master_clf = dict()

for slot in slot_names:
    print("Training for slot =", slot)
    clf = LogisticRegression(random_state=42)
    y_train = train_df["slots"].apply(lambda slots: transform_to_binary(slots, slot))
    # print(y_train.shape)
    print()
    clf.fit(x_train_labse, y_train)
    print(f"Train score = {clf.score(x_train_labse, y_train) * 100:.2f} %")
    master_clf[slot] = clf
    print("------------")

print("Completed")

# create predicted dictionary for each item
train_df["pred_slots"] = train_df["text"].apply(lambda x: dict())
print(train_df.iloc[0]["pred_slots"])
train_df.head()

# double check train scores
for slot in slot_names:
    y_pred = master_clf[slot].predict(x_train_labse)
    # print("Train score =", master_clf[slot].score(x_train_tfidf, y_train))
    y_train = train_df["slots"].apply(lambda slots: transform_to_binary(slots, slot))
    print("Accuracy score =", accuracy_score(y_train, y_pred) * 100) # should be the same

    # go through non-empty result and add to pred_slots
    for i, item in enumerate(y_pred):
        # print(i)
        if item is not None and item != "":
            item_slot = train_df.iloc[i]["pred_slots"]
            # print(item_slot)
            item_slot.update({slot:item})
            train_df.at[i, "pred_slots"] = item_slot

train_df.head()

# model_name = "consciousAI/question-answering-roberta-base-s-v2"
model_name = "deepset/roberta-base-squad2"

def get_prediction(question, context, model=model_name):
    
    nlp = pipeline('question-answering', model=model_name, tokenizer=model_name)
    
    QA_input = {
        'question': question,
        'context': context
    }
    answer = nlp(QA_input)

    final_answer = answer["answer"].lower()
    
    return final_answer

# Slot:Question dictionary:
SQ_dict = {
    "hotel-name" : "What is the hotel name?",
    "restaurant-food" : "What kind of food does the restaurant serve?",
    "restaurant-name" : "What is the name of the restaurant?",
}

def update_pred_slots(row):
    pred_slots = row["pred_slots"]
    for slot, query in SQ_dict.items():
        if slot in pred_slots:
            pred_slots[slot] = get_prediction(query, row["text"])
    return pred_slots

train_df["pred_slots"] = train_df.apply(update_pred_slots, axis=1)

train_df.head()

# get accuracy on train set
def get_accuracy(gold_slots, pred_slots, slot_lists=slot_names):
    """
        return accuracy of predicted slots vs gold slots in dictionary form
    """
    correct_count = 0
    for gold_slot, pred_slot in zip(gold_slots, pred_slots):
        gold = {k:v 
                for k, v in gold_slot.items()
                if k in slot_names
               }
        # print(gold)
        sys = {k:v 
                for k, v in pred_slot.items()
                if k in slot_names
               }
        # print(sys)
        # if gold_slot == pred_slot:
        if gold == sys:
            correct_count += 1
    return correct_count / len(gold_slots)

get_accuracy(train_df["slots"], train_df["pred_slots"])

# test on dev set
dev_filename = "./data/dev.csv"
dev_df = pd.read_csv(dev_filename)
print(dev_df.columns)
dev_df.head()

# filter to hotel only
# category = 'find_hotel'
# dev_df = dev_df.query("intent==@category")
print(len(dev_df))
# dev_df.reset_index(inplace=True)
# dev_df.drop(columns=["index"], inplace=True)
dev_df.head()

# apply some pre-processing
dev_df["text_lower"] = dev_df["text"].str.lower()
# convert slots to json objects
dev_df["slots"] = dev_df["slots"].apply(lambda s: eval(s))
dev_df.head()

x_dev = dev_df["text"]
# x_dev_tfidf = vectorizer.transform(x_dev)
x_dev = dev_df["text_lower"]
x_dev = list(x_dev)
# x_train_tfidf = vectorizer.fit_transform(x_train)


x_dev_labse = model.encode(x_dev)

print(x_dev_labse.shape)

# create predicted dictionary for each item
dev_df["pred_slots"] = dev_df["text"].apply(lambda x: dict())
dev_df.head()

for slot in slot_names:
    # y_dev = dev_df["target"]
    y_dev = dev_df["slots"].apply(lambda slots: transform_to_binary(slots, slot))
    
    dev_pred = master_clf[slot].predict(x_dev_labse)
    # dev_score = clf.score(x_dev_tfidf, y_dev)
    print("Slot prediction =", slot)
    print(f"Dev accuracy = {accuracy_score(y_dev, dev_pred) * 100:.2f} %")

    # go through non-empty result and add to pred_slots
    for i, item in enumerate(dev_pred):
    # print(i)
        if item is not None and item != "":
            item_slot = dev_df.iloc[i]["pred_slots"]
            # print(item_slot)
            item_slot.update({slot:item})
            dev_df.at[i, "pred_slots"] = item_slot
            
    # print("Dev accuracy =", dev_score * 100)

dev_df.head()

def update_pred_slots(row):
    pred_slots = row["pred_slots"]
    for slot, query in SQ_dict.items():
        if slot in pred_slots:
            pred_slots[slot] = get_prediction(query, row["text"])
    return pred_slots

dev_df["pred_slots"] = dev_df.apply(update_pred_slots, axis=1)

dev_df

# calculate overall accuracy 
score = get_accuracy(dev_df["slots"], dev_df["pred_slots"])
print(f"Overall accuracy = {score:.2f} %")

# output
dev_df.to_csv("./data/dev_step3.csv")

#test set
test_filename = "./data/test.csv"
test_df = pd.read_csv(test_filename)


x_test = test_df["text_lower"]
x_test = list(x_test)

x_test_labse = model.encode(x_test)

print(x_test_labse.shape)

# create predicted dictionary for each item
test_df["pred_slots"] = test_df["text"].apply(lambda x: dict())


for slot in slot_names:
    
    test_pred = master_clf[slot].predict(x_test_labse)

    # go through non-empty result and add to pred_slots
    for i, item in enumerate(test_pred):
        if item is not None and item != "":
            item_slot = test_df.iloc[i]["pred_slots"]
            # print(item_slot)
            item_slot.update({slot:item})
            test_df.at[i, "pred_slots"] = item_slot

test_df["pred_slots"] = test_df.apply(update_pred_slots, axis=1)

# output
test_df.to_csv("./data/test_step3.csv")
