==> 08_Consolidation.ipynb <==
# Step 4 Consolidate all outputs from different models
This notebook consolidates the output from `intent classification model`, `8/11 Slots classification model`, `3/11 Slot classification model`.

import pandas as pd
from sklearn.metrics import accuracy_score

dev_files = ["./data/dev.csv","./data/dev_step1.csv", "./data/dev_step2.csv", "./data/dev_step3.csv"]

dev0_df = pd.read_csv(dev_files[0])# processed original dev set
dev1_df = pd.read_csv(dev_files[1],index_col=0)# contains intent prediction
dev2_df = pd.read_csv(dev_files[2],index_col=0)# contains 8/11 slots prediction
dev3_df = pd.read_csv(dev_files[3],index_col=0)# contains 3/11 slots prediction

print(dev0_df.shape)
print(dev1_df.shape)
print(dev2_df.shape)
print(dev3_df.shape)

test_files = ["./data/test.csv","./data/test_step1.csv", "./data/test_step2.csv", "./data/test_step3.csv"]

test0_df = pd.read_csv(test_files[0])# processed original test set
test1_df = pd.read_csv(test_files[1],index_col=0)# contains intent prediction
test2_df = pd.read_csv(test_files[2],index_col=0)# contains 8/11 slots prediction
test3_df = pd.read_csv(test_files[3],index_col=0)# contains 3/11 slots prediction

print(test0_df.shape)
print(test1_df.shape)
print(test2_df.shape)
print(test3_df.shape)

dev0_df#["answer_raw"]

# preprocessing before merge:
def preprocess_n_consolidation(df1, df2, df3):
    """
    pass in either (dev1_df, dev2_df, dev3_df) 
    or (test1_df, test2_df, test3_df) for consolidation
    """
    order = [
        "hotel-area","hotel-internet","hotel-name","hotel-parking","hotel-pricerange",'hotel-stars',"hotel-type",
        "restaurant-area","restaurant-food","restaurant-name","restaurant-pricerange",
    ]
    
    df1 = df1.rename(columns={"predicted":"pred_intent"})
    df2 = df2.rename(columns={"pred_slots":"pred_slots_2"})
    df3 = df3.rename(columns={"pred_slots":"pred_slots_3"})
    
    df = df1.merge(df2[["pred_slots_2"]], left_index=True, right_index=True, how="left",validate="1:1")
    df = df.merge(df3[["pred_slots_3"]], left_index=True, right_index=True, how="left",validate="1:1")
    
    df['pred_slots_2'] = df['pred_slots_2'].apply(lambda x: eval(x))
    df['pred_slots_3'] = df['pred_slots_3'].apply(lambda x: eval(x))
    
    def Merge(dict1, dict2):
        res = {**dict1, **dict2}
        return res

    df['pred_slots'] = df.apply(lambda x: Merge(x.pred_slots_2, x.pred_slots_3), axis=1)

    def reorder_dict(dict, order):
        reordered_dict = {key: dict[key] for key in order if key in dict.keys()}
        return reordered_dict

    df['pred_slots'] = df.apply(lambda x: reorder_dict(x.pred_slots, order), axis=1)

    def dict_2_list(dict):
        l = []
        for k, v in dict.items():
            l.append(k +"="+v)
        return l
    
    df['pred_answers'] = df.apply(lambda x: dict_2_list(x.pred_slots), axis=1)

    def to_answer_raw(pred_intent, pred_answers):
        pred_answers.insert(0, pred_intent)
        s = "|".join(pred_answers)
        return s

    df['pred_answer_raw'] = df.apply(lambda x: to_answer_raw(x.pred_intent, x.pred_answers), axis=1)
    
    return(df)

dev4_df = preprocess_n_consolidation(dev1_df, dev2_df, dev3_df)
# dev4_df

test4_df = preprocess_n_consolidation(test1_df, test2_df, test3_df)
# test4_df

slot_names = ['hotel-name', 'hotel-stars', 
              'hotel-area', 'hotel-internet', 
              'hotel-pricerange', 'hotel-parking', 
              'hotel-type',
              'restaurant-food', 
              'restaurant-name', 
              'restaurant-pricerange', 
              'restaurant-area']

def get_accuracy(gold_slots, pred_slots, slot_lists=slot_names):
    """
        return accuracy of predicted slots vs gold slots in dictionary form
    """
    correct_count = 0
    for gold_slot, pred_slot in zip(gold_slots, pred_slots):
        gold = {k:v 
                for k, v in gold_slot.items()
                if k in slot_names
               }
        # print(gold)
        sys = {k:v 
                for k, v in pred_slot.items()
                if k in slot_names
               }
        # print(sys)
        # if gold_slot == pred_slot:
        if gold == sys:
            correct_count += 1
    return correct_count / len(gold_slots)



#dev
get_accuracy(dev0_df["slots"].apply(lambda x: eval(x)), dev4_df["pred_slots"])

#dev
print("Accuracy score =", accuracy_score(dev0_df["answer_raw"], dev4_df["pred_answer_raw"]) * 100)

# Output
dev4_df.to_csv("./data/dev_step4.csv")
dev4_df[["text","pred_answer_raw"]].to_csv('./data/dev_pred.txt', sep='\t', index=False)

# Output test
test4_df.to_csv("./data/test_step4.csv")
test4_df[["text","pred_answer_raw"]].to_csv('./data/test_pred.txt', sep='\t', index=False)



test4_df

kaggle_df = test4_df.copy()
kaggle_df = kaggle_df.reset_index()


kaggle_df = kaggle_df.rename(columns={"index":"ID", "pred_answer_raw":"Expected"})

kaggle_df[["ID", "Expected"]].to_csv('./data/WOZ_test_ans.csv', index=False)

kaggle_df[kaggle_df["Expected"].str.contains("restaurant-name")]
